requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.2
Pillow>=10.2.0
python-dotenv>=1.0.1 